% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetaAQE.R
\name{getMetaAQE}
\alias{getMetaAQE}
\title{getMetaAQE}
\usage{
getMetaAQE(borough_sf = NULL)
}
\arguments{
\item{borough_sf}{A simple-features data frame containing London Borough polygons. , Default: NULL}
}
\value{
A data frame containing meta data for all AQE monitoring sites in London
}
\description{
Get the meta data for all Air Quality England (AQE) monitoring sites in London
}
\details{
If borough_sf is not provided local authority name and inner/outer London category will not be returned.
London Borough boundary data is availabe from the \href{https://data.london.gov.uk/dataset/statistical-gis-boundary-files-london}{London DataStore}
and \href{https://www.ordnancesurvey.co.uk/business-and-government/products/boundaryline.html}{Ordnance Survey}.
}
\examples{
\dontrun{
if(interactive()){
# Without borough_sf

meta_data <- getMetaAQE()

# With borough_sf
boroughs <- st_read("London_Borough.shp")
meta_data <- getMetaAQE(borough_sf = boroughs)
 }
}
}
