% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetaBreathe.R
\name{getMetaBreathe}
\alias{getMetaBreathe}
\title{getMetaBreathe}
\usage{
getMetaBreathe(borough_sf = NULL)
}
\arguments{
\item{borough_sf}{A simple-features data frame containing London Borough polygons. , Default: NULL}
}
\value{
A data frame containing meta data for all Breathe London monitoring sites in London
}
\description{
Get the meta data for all Breathe London monitoring sites in London
}
\details{
If borough_sf is not provided local authority name and inner/outer London category will not be returned.
London Borough boundary data is availabe from the \href{https://data.london.gov.uk/dataset/statistical-gis-boundary-files-london}{London DataStore}
and \href{https://www.ordnancesurvey.co.uk/business-and-government/products/boundaryline.html}{Ordnance Survey}.
}
\examples{
\dontrun{
if(interactive()){
# Without borough_sf

meta_data <- getMetaBreathe()

# With borough_sf
boroughs <- st_read("London_Borough.shp")
meta_data <- getMetaBreathe(borough_sf = boroughs)
 }
}
}
\seealso{
\code{\link[utils]{read.table}}
 \code{\link[lubridate]{ymd_hms}}
}
