% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importAllLondon.R
\name{importAllLondon}
\alias{importAllLondon}
\title{importAllLondon}
\usage{
importAllLondon(start_date = Sys.Date() - 1, end_date = Sys.Date() - 1,
  sites = "all", species = "all", borough_sf = NULL,
  meta_data = NULL, verbose = TRUE)
}
\arguments{
\item{start_date}{Inclusive start date of data request. Must be a date object, Default: Sys.Date() - 1}

\item{end_date}{Inclusive end date of data request. Must be date object, Default: Sys.Date() - 1}

\item{sites}{character vector of site codes or 'all' to fetch all available sites, Default: 'all'}

\item{species}{character vector of species codes or 'all' to fetch all available species, Default: 'all'}

\item{borough_sf}{A simple-features data frame containing London Borough polygons, Default: NULL}

\item{meta_data}{Meta data dataframe, as fetched by importAllLondon etc. If not provided will be fetched, Default: NULL}

\item{verbose}{logical. Include site meta data in returned data frame, Default: TRUE}
}
\value{
A data frame of hourly monitoring results
}
\description{
Import air quality data from across London for specified parameters
}
\details{
If borough_sf is not provided, local authority name and inner/outer london category will not be returned.
}
\examples{
\dontrun{
if(interactive()){
# For all sites/species:
df <- importAllLondon(start_date = as.Date('2019-01-01'),
end_date = as.Date('2019-01-07'))

# For specified sites and species:
df <- importAllLondon(start_date = as.Date('2019-01-01'),
end_date = as.Date('2019-01-07'),
sites = c("BG1", "WL1", "MY1"),
species = c("NO2", "PM10", "PM2.5"))

# Providing meta data
meta_data <- getMetaAllLondon()
df <- importAllLondon(start_date = as.Date('2019-01-01'),
end_date = as.Date('2019-01-07'),
sites = c("BG1", "WL1", "MY1"),
species = c("NO2", "PM10", "PM2.5"),
meta_data = meta_data)
 }
}
}
\seealso{

}
