% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importBreathe.R
\name{importBreathe}
\alias{importBreathe}
\title{importBreathe}
\usage{
importBreathe(start_date = Sys.Date() - 1, end_date = Sys.Date() - 1,
  sites = "all", species = "all", borough_sf = NULL,
  meta_data = NULL, verbose = TRUE)
}
\arguments{
\item{start_date}{Inclusive start date of data request. Must be a date object, Default: Sys.Date() - 1}

\item{end_date}{Inclusive end date of data request. Must be date object, Default: Sys.Date() - 1}

\item{sites}{character vector of site codes or 'all' to fetch all available sites, Default: 'all'}

\item{species}{character vector of species codes or 'all' to fetch all available species, Default: 'all'}

\item{borough_sf}{A simple-features data frame containing London Borough polygons. , Default: NULL}

\item{meta_data}{Meta data dataframe, as fetched by importLondon etc. If not provided will be fetched, Default: NULL}

\item{verbose}{logical. Include site meta data in returned data frame, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Import air quality data from across the Breathe London Network for specified parameters
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
# For all sites/species:
df <- importBreathe(start_date = as.Date('2019-01-01'),
end_date = as.Date('2019-01-07'))

# For specified sites and species (only NO2 available for Breathe London):
df <- importBreathe(start_date = as.Date('2019-01-01'),
end_date = as.Date('2019-01-07'),
sites = c("2245", "16245"),
species = "NO2")

# Providing meta data
meta_data <- getMetaLondon()
df <- importLondon(start_date = as.Date('2019-01-01'),
end_date = as.Date('2019-01-07'),
sites = c(c("GB6", "CR5")),
species = c("NO2", "PM10", "PM2.5"),
meta_data = meta_data)
 }
}
}
\seealso{
\code{\link[utils]{read.table}}
}
